v2.0.767.39141

+++ tools/EVA.SDK.Generator.V2/Commands/Generate/Outputs/OutputWriter.cs
@@ -9,7 +9,13 @@ internal class OutputWriter
   private readonly string _directory;
   private readonly ILogger _logger;
 
-  private readonly Dictionary<string, long> _writtenFiles = new();
+  private readonly Dictionary<string, (Result result, long size)> _writtenFiles = new();
+
+  internal enum Result
+  {
+    Overwritten,
+    Kept
+  }
 
   internal OutputWriter(string directory, ILogger logger)
   {
@@ -50,8 +56,17 @@ internal class OutputWriter
 
     var path = Path.Combine(_directory, file);
     EnsureDirectoryExists(path);
-    await File.WriteAllTextAsync(path, content);
-    _writtenFiles.Add(file, new FileInfo(path).Length);
+
+    // Only overwrite if file doesn't exist or content is different
+    if (!File.Exists(path) || await File.ReadAllTextAsync(path) != content)
+    {
+      await File.WriteAllTextAsync(path, content);
+      _writtenFiles.Add(file, (Result.Overwritten, new FileInfo(path).Length));
+    }
+    else
+    {
+      _writtenFiles.Add(file, (Result.Kept, new FileInfo(path).Length));
+    }
   }
 
   internal DisposableCallback<Stream> WriteStreamAsync(string file)
@@ -62,32 +77,37 @@ internal class OutputWriter
     EnsureDirectoryExists(path);
     return new DisposableCallback<Stream>(File.OpenWrite(path), () =>
     {
-      _writtenFiles.Add(file, new FileInfo(path).Length);
+      _writtenFiles.Add(file, (Result.Overwritten, new FileInfo(path).Length));
     });
   }
 
-  internal string ToReport()
+  internal void DeleteRemaining()
+  {
+    var allWritten = _writtenFiles.Keys.Select(x => Path.GetFullPath(x, _directory)).ToHashSet();
+
+    foreach(var x in Directory.EnumerateFiles(_directory, "*", SearchOption.AllDirectories))
+    {
+      if (!allWritten.Contains(x))
+      {
+        File.Delete(x);
+      }
+    }
+  }
+
+  internal void WriteReport(ILogger logger)
   {
     var sb = new StringBuilder();
 
-    var totalSize = _writtenFiles.Values.Sum();
-    sb.Append("Wrote ").Append(_writtenFiles.Count).Append(" files with total size of ").Append(StringHelpers.FormatSize(totalSize));
-    if (_writtenFiles.Count is <= 1 or > 40) return sb.ToString();
+    var keptFiles = _writtenFiles.Values.Count(x => x.result == Result.Kept);
+    logger.LogInformation($"No change to {keptFiles} files");
 
-    var records = _writtenFiles
-      .Select(f => (name: f.Key, sizeStr: StringHelpers.FormatSize(f.Value), size: f.Value))
-      .OrderByDescending(x => x.size)
+    var changedFiles = _writtenFiles
+      .Where(x => x.Value.result == Result.Overwritten)
+      .Select(x => (name: x.Key, x.Value.result, x.Value.size))
       .ToList();
-    var sizeColumnWidth = records.Max(x => x.sizeStr.Length);
 
-    sb.AppendLine(":");
-
-    foreach (var f in records)
-    {
-      sb.Append("  ").Append(f.sizeStr.PadLeft(sizeColumnWidth)).Append(' ').Append(((int)(100.0f * f.size / totalSize)).ToString().PadLeft(3)).Append("% ").AppendLine(f.name);
-    }
-
-    return sb.ToString();
+    var totalSize = changedFiles.Sum(x => x.size);
+    logger.LogInformation($"Wrote {changedFiles.Count} files with total size of {StringHelpers.FormatSize(totalSize)}");
   }
 
   internal class DisposableCallback<T> : IAsyncDisposable where T : IAsyncDisposable
